fix MENU_HIDE

+++ usr/lib/archboot/update/gnome.sh
@@ -19,9 +19,10 @@ EOF
     _HIDE_MENU="avahi-discover bssh bvnc org.gnome.Extensions org.gnome.FileRoller org.gnome.gThumb org.gnome.gedit fluid vncviewer lstopo qvidcap qv4l2"
     echo "Hiding ${_HIDE_MENU} menu entries..."
     for i in ${_HIDE_MENU}; do
-        [[ -f /usr/share/applications/"${i}".desktop ]] || break
-        echo "[DESKTOP ENTRY]" > /usr/share/applications/"${i}".desktop
-        echo 'NoDisplay=true' >> /usr/share/applications/"${i}".desktop
+        if [[ -f /usr/share/applications/"${i}".desktop ]]; then
+            echo "[DESKTOP ENTRY]" > /usr/share/applications/"${i}".desktop
+            echo 'NoDisplay=true' >> /usr/share/applications/"${i}".desktop
+        fi
     done
 }
